+++ docs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # GoMLX changelog
 
+# v0.19.4: 2024/05/24 added Vector Neural Networks (VNNs)
+
+* Vector Neural Networks (VNN): allows one to build 3D rotation (SO(3)) equivariant and/or invariant networks. See package `ml/layers/vnn`.
+* Package `xla`
+  * Remove dependencies to `gopjrt` internal protos: requires updated `Gopjrt`.
+* Package `tensors`
+  * Fixed pretty-print of booleans.
+
 # v0.19.3: 2024/05/20 Many SimpleGo improvements. 
 
 * v0.19.2 skipped ... issues with the release.